serial_service/requirements.txt
pyserial==3.5
websockets==15.0.1
fastmcp==2.5.2
fastapi==0.115.12
fastapi-cli==0.0.7
requests==2.32.3
pika==1.3.2